CHANGELOG-827-2014-1204 Removed all magic numbers.  - Constants placed in their respective classes; global constants in Constants file Changed all ints to NSUInteger. Changed all floats to CGFloat. Changed object positioning to follow Screen bounds.  - Fixed bug with Background View not scaling to the screen properly Fixed "snapping" of Cloud View between animation cycles

+++ Air Raid/KLBBattleViewController.h
@@ -8,9 +8,9 @@
 
 #import <UIKit/UIKit.h>
 
-extern float const KLB_TOUCH_BOTTOM_BUFFER_PERCENTAGE;
-extern float const KLB_MINIMUM_PRESS_DURATION;
-extern float const KLB_ALLOWABLE_MOVEMENT;
+extern CGFloat const KLB_TOUCH_BOTTOM_BUFFER_PERCENTAGE;
+extern CGFloat const KLB_MINIMUM_PRESS_DURATION;
+extern CGFloat const KLB_ALLOWABLE_MOVEMENT;
 
 @interface KLBBattleViewController : UIViewController
 @property (retain, nonatomic) IBOutlet UILabel *scorePlaceholderLabel;
